convert topk_weights to FP32 in kernel

+++ python/sglang/srt/layers/quantization/fp8.py
@@ -1002,9 +1002,7 @@ class Fp8MoEMethod:
                 x,
                 layer.w13_weight,
                 layer.w2_weight,
-                topk_weights.to(
-                    torch.float
-                ),  # See [Note] topk_weights should be float32 for now.
+                topk_weights,
                 topk_ids,
                 False,  # inplace See [Note] inplace should be False in fused_experts.
                 False,  # use_int8_w8a8
