+++ python/sglang/srt/layers/quantization/w8a8_int8.py
@@ -290,9 +290,7 @@ class W8A8Int8MoEMethod:
                 x,
                 layer.w13_weight,
                 layer.w2_weight,
-                topk_weights.to(
-                    torch.float
-                ),  # See [Note] topk_weights should be float32 for now.
+                topk_weights,
                 topk_ids,
                 False,  # inplace See [Note] inplace should be False in fused_experts.
                 True,  # use_int8_w8a8
